requirements.txt
absl-py @ file:///croot/absl-py_1714140470852/work
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1720453481653/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
contourpy==1.1.1
cycler==0.12.1
filelock @ file:///croot/filelock_1700591183607/work
fonttools==4.53.1
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
grpcio @ file:///croot/grpc-split_1716834572340/work
idna @ file:///croot/idna_1714398848350/work
importlib-metadata @ file:///croot/importlib_metadata-suite_1704813515092/work
importlib_resources==6.4.2
Jinja2 @ file:///croot/jinja2_1716993405101/work
joblib==1.4.2
kiwisolver==1.4.5
Markdown @ file:///croot/markdown_1671541909495/work
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.7.5
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
networkx @ file:///croot/networkx_1690561992265/work
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
opencv-python==4.10.0.84
packaging==24.1
pandas==2.0.3
pillow @ file:///croot/pillow_1721059439630/work
protobuf==4.25.3
pyparsing==3.1.2
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML @ file:///croot/pyyaml_1698096049011/work
requests @ file:///croot/requests_1721410876868/work
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.11.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
statannot==0.2.3
sympy @ file:///croot/sympy_1701397643339/work
tensorboard @ file:///home/conda/feedstock_root/build_artifacts/tensorboard_1718113407052/work/tensorboard-2.17.0-py3-none-any.whl#sha256=859a499a9b1fb68a058858964486627100b71fcb21646861c61d31846a6478fb
tensorboard-data-server @ file:///croot/tensorboard-data-server_1681498183723/work/tensorboard_data_server-0.7.0-py3-none-manylinux2014_x86_64.whl
threadpoolctl==3.5.0
torch==2.1.1
torchvision @ file:///workspace/croot/torchvision_1693425165028/work
tqdm==4.66.5
triton==2.1.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.1
urllib3 @ file:///croot/urllib3_1718912636303/work
Werkzeug @ file:///croot/werkzeug_1716533313107/work
zipp @ file:///croot/zipp_1704206909481/work
